Neovim config (using lua!)
I've been using nvim for a while now, but have just been symlinking my
vimrc to neovim. I went all-in, converting to lua, and using a lot of
lua plugins as well. I still need to clean up the files a bit; they're a
bit of a mess. TBH they're a clusterfuck, actually